requirements.txt
gradio>=4.0.0
google-generativeai==0.8.3
sentence-transformers==3.2.1
numpy==2.1.3
pandas==1.5.3
plotly==5.16.1
pypdf2==3.0.1
openpyxl==3.1.5
Flask==2.3.2
requests==2.31.0
google-genai==1.52.0
Jinja2==3.1.2
